nwe co mmit

+++ src/Components/dashboard.js
@@ -5,9 +5,11 @@ import axios from 'axios'
    const [userDetails, setUserDetails] = useState([])
 
    const LogOut = async()=>{
-    alert('logging out')
-    const clear =  sessionStorage.clear()
-    if(clear) window.location.assign('https://hg-pre-interview-cd43cd.netlify.app/login')
+    const logUserOut = await axios.get('https://hg-pre-interview-task-1000.herokuapp.com/logout')
+    if(logUserOut.data === "success"){
+      sessionStorage.clear()
+      window.location.assign('https://hg-pre-interview-cd43cd.netlify.app/login')
+    }
   }
    useEffect(()=>{
     // alert('hello there i am trying to control the width of the screen')
